Moved db export endpoint to slightly more logical app (call_centre).
