+++ cla_backend/apps/call_centre/sql/ExportStandardTables.sql
