+++ cla_backend/apps/call_centre/urls.py
@@ -45,4 +45,7 @@ urlpatterns = patterns('',
     url(r'^', include(case_one2many_router.urls)),
     url(r'^', include(router.urls)),
     url(r'^', include(timer_router.urls)),
+    url(r'^dbexport/$',
+        views.DBExportView.as_view(),
+        name='db_export')
 )
